+++ android/src/main/java/io/openim/flutter_openim_sdk/manager/ConversationManager.java
@@ -162,4 +162,13 @@ public class ConversationManager extends BaseManager {
                 int2long(methodCall, "status")
         );
     }
+
+    public void setOneConversationBurnDuration(MethodCall methodCall, MethodChannel.Result result) {
+        Open_im_sdk.setOneConversationBurnDuration(
+                new OnBaseListener(result, methodCall),
+                value(methodCall, "operationID"),
+                value(methodCall, "conversationID"),
+                value(methodCall, "burnDuration")
+        );
+    }
 }